backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
twilio
pydantic
numpy
soundfile
aiohttp